+++ tests/test_mananging_actor.py
@@ -5,7 +5,6 @@
 
 import asyncio
 import heapq
-import logging
 from dataclasses import dataclass, replace
 from datetime import datetime, timedelta, timezone
 from typing import AsyncIterator, Callable, Iterator, cast
@@ -70,13 +69,27 @@ class MockActor(Actor):
     ) -> None:
         """Initialize the actor."""
         super().__init__(name="MockActor")
-        logging.info("MockActor created")
         self.initial_dispatch = initial_dispatch
         self.receiver = receiver
 
     async def _run(self) -> None:
         while True:
             await asyncio.sleep(1)
+
+    @classmethod
+    async def create(
+        cls, initial_dispatch: DispatchInfo, receiver: Receiver[DispatchInfo]
+    ) -> "MockActor":
+        """Create a new actor."""
+        actor = cls(initial_dispatch, receiver)
+        return actor
+
+    @classmethod
+    async def create_fail(
+        cls, __: DispatchInfo, _: Receiver[DispatchInfo]
+    ) -> "MockActor":
+        """Create a new actor."""
+        raise ValueError("Failed to create actor")
 
 
 @dataclass
@@ -101,7 +114,7 @@ async def test_env() -> AsyncIterator[TestEnv]:
     channel = Broadcast[Dispatch](name="dispatch ready test channel")
 
     actors_service = ActorDispatcher(
-        actor_factory=MockActor,
+        actor_factory=MockActor.create,
         running_status_receiver=channel.new_receiver(),
         dispatch_identity=lambda dispatch: dispatch.id,
     )
@@ -145,14 +158,11 @@ async def test_simple_start_stop(
     fake_time.shift(timedelta(seconds=1))
     await asyncio.sleep(1)
     await asyncio.sleep(1)
-    logging.info("Sent dispatch")
 
     event = test_env.actor(1).initial_dispatch
     assert event.options == {"test": True}
     assert event.components == dispatch.target
     assert event.dry_run is False
-
-    logging.info("Received dispatch")
 
     assert test_env.actor(1).is_running is True
 
@@ -165,6 +175,47 @@ async def test_simple_start_stop(
     # pylint: disable=protected-access
     assert 1 not in test_env.actors_service._actors
     # pylint: enable=protected-access
+
+
+async def test_start_failed(
+    test_env: TestEnv, fake_time: time_machine.Coordinates
+) -> None:
+    """Test auto-retry after 60 seconds."""
+    # pylint: disable=protected-access
+    test_env.actors_service._actor_factory = MockActor.create_fail
+
+    now = _now()
+    duration = timedelta(minutes=10)
+    dispatch = test_env.generator.generate_dispatch()
+    dispatch = replace(
+        dispatch,
+        id=1,
+        active=True,
+        dry_run=False,
+        duration=duration,
+        start_time=now,
+        payload={"test": True},
+        type="UNIT_TEST",
+        recurrence=replace(
+            dispatch.recurrence,
+            frequency=Frequency.UNSPECIFIED,
+        ),
+    )
+
+    # Send status update to start actor, expect no DispatchInfo for the start
+    await test_env.running_status_sender.send(Dispatch(dispatch))
+    fake_time.shift(timedelta(seconds=1))
+
+    # Replace failing mock actor factory with a working one
+    test_env.actors_service._actor_factory = MockActor.create
+
+    # Give retry task time to start
+    await asyncio.sleep(1)
+
+    fake_time.shift(timedelta(seconds=65))
+    await asyncio.sleep(65)
+
+    assert test_env.actor(1).is_running is True
 
 
 def test_heapq_dispatch_compare(test_env: TestEnv) -> None:
@@ -258,7 +309,7 @@ async def test_manage_abstraction(
     generator: DispatchGenerator,
     strategy: MergeStrategy | None,
 ) -> None:
-    """Test Dispatcher.start_dispatching sets up correctly."""
+    """Test Dispatcher.start_managing sets up correctly."""
     identity: Callable[[Dispatch], int] = (
         strategy.identity if strategy else lambda dispatch: dispatch.id
     )
@@ -284,7 +335,10 @@ async def test_manage_abstraction(
         sender = channel.new_sender()
 
         async def new_mock_receiver(
-            _: Dispatcher, dispatch_type: str, *, merge_strategy: MergeStrategy | None
+            _: Dispatcher,
+            dispatch_type: str,
+            *,
+            merge_strategy: MergeStrategy | None,
         ) -> Receiver[Dispatch]:
             assert dispatch_type == "MANAGE_TEST"
             assert merge_strategy is strategy
@@ -294,16 +348,18 @@ async def test_manage_abstraction(
             "frequenz.dispatch._dispatcher.Dispatcher.new_running_state_event_receiver",
             new_mock_receiver,
         ):
-            await dispatcher.start_dispatching(
+            await dispatcher.start_managing(
                 dispatch_type="MANAGE_TEST",
-                actor_factory=MockActor,
+                actor_factory=MockActor.create,
                 merge_strategy=strategy,
             )
 
         # pylint: disable=protected-access
         assert "MANAGE_TEST" in dispatcher._actor_dispatchers
         actor_manager = dispatcher._actor_dispatchers["MANAGE_TEST"]
-        assert actor_manager._actor_factory == MockActor
+        # pylint: disable=comparison-with-callable
+        assert actor_manager._actor_factory == MockActor.create
+        # pylint: enable=comparison-with-callable
 
         dispatch = Dispatch(
             replace(